add passport auth and user/role authorization

+++ server/models/catModel.js
@@ -38,10 +38,10 @@ const addCat = async (cat, res) => {
   }
 };
 
-const deleteCatById = async (catId, res) => {
+const deleteCatById = async (catId, owner, res) => {
   try {
     const [rows] = await promisePool.
-    query("DELETE FROM wop_cat WHERE wop_cat.cat_id = ?", [catId]);
+    query("DELETE FROM wop_cat WHERE wop_cat.cat_id = ? AND wop_cat.owner = ?", [catId, owner]);
     return rows;
   } catch (e) {
     console.error("error", e.message);
